Expanded store reducers and types into own files
