+++ src/store/index.js
@@ -3,8 +3,6 @@
 import * as Redux from 'redux'
 import {createStore} from 'redux'
 
-import {
-  EditorReducer
-} from './reducers'
+import EditorReducer from './reducers/EditorReducer'
 
 export const store: Redux.Store = createStore(EditorReducer)
